Multiple slots implemented, more forward movement when attacking
Players now have multiple slots that they can swap with the d-pad.

Player now move forward when attacking more, based on the "movementSpeedMultiplier" float inside the "Weapon Item" script, this multiplier is applied to the "currentSpeedMultiplier" inside the PlayerManager.

Solved bools disablePlayerInteraction and isAttacking setting to false when chaining attacks by checking if an attack was chained in the "Reset Combo" script and setting them to true if so.