Input queue
Added the ability to queue input like pressing or releasing buttons.

Input system rework.

SetPlayerManagerBool script is now only used on enter state and should be that way as the on exit state is unreliable as it is not always called.
For example setting bools for isAttacking, isRolling... to false is now done in the enter state of the empty state in the override layer.