Slot change animation
Now the players does a small animation when it is changing items.

Added the flag disablePlayerActions that stops the player from dodging, attacking, blocking...

This flag only stops actions but not movement like the isPlayerStuckInAnimation does.

if the isPlayerStuckInAnimation is true the movement floats in the animator get set to 0.

Changed the strafe right/left animations.

Added kite shield.