Blocking now reduces damage
While isBlocking is true, damage from being hit is now reduced based on block item absorption %.

Blocking an attack now stops the player for a small amount of time from moving.

Added a currentItemType in the inventory manager that is set when an item is loaded.

This allow for example to avoid doing the type check on the item trying to attack with, if it is not of the right type (weapon) then the attack will not go through.