Player color at spawn and hitbox bug fix
Now the player gets assigned a random color on spawn.

Comboing an attack now closes the collider to be opened again in the animation event (the combo made the closing event not run), this allows the hit list of the collider to be emptied and for the successive hit of the combo to register as it previously sometimes didn't.

The collider is closed in the ResetCombo script that gets called on the exit from the attack animations (whether the player combos or not (if the player does not combo the collider is simply closed twice, here and in the animation event, it is not a problem)).