ServerRpcs refactor
Moved most of the player rpcs to the network manager.

Changed the way actions are passed through rpcs, instead of simply passing the attack animation through rpc for example, the whole Attack method is now passed, the same applies for all the actions such as rolls, backdashes and so on.

Temporarily disabeld collision with static (weapon bounce on walls).