Player distancing, course sync and iframes
Player now have a capsule collider that distances them from other players (it only collides with other instances of the same collider).

Lifts and cannons are now synced between player, also lifts now move the players that stand on them with the lift movement.

Added areIFramesActive bool that gets set to true (and false) when rolling or backdashing, this allows players to avoid being damaged.