Animation system rework
Game now uses animancer to manage it's animations.

Every player has a PlayerAnimationsDatabase that stores all the animations that it can do.

Removed chance to attack with left hand or block with right hand because of runtime mirroring of the animation limitation of the animancer system (it does not allow runtime mirroring).

The system could be implemented again simply adding mirrored version of each animation in the animation database, but is for now removed.