More animations rework and inventory system changes
More animation logic has been changed to the Animancer system.

The inventory now uses a PlayerItemDatabase that has all the items and a corresponding id, the player inventory slots are now using the id to specify which item they possess.

This allows ServerRpc to specify what item to equip via the id.