Attacks system rework and collision changes
The player now has a capsule collider around set as the hurtbox to improve hit detection reliability.

Changed the way the moveset is assigned to a weapon.

Now each weapon has a moveset scriptable object and each moveset has a series of attack moves (another scriptable object) that define things like animation name, speed, damage multipliers and so on...