Attack deflection, player rotation enabled only at attack start
Shields now have a deflection value that ranges from 1 to 4, attacks have the same, if a shield blocks an attack with this deflection value < than the shield, then the attack will bounce off the shield.

Player is now only allowed to rotate at the start of an attack, for how much it's defined in each attack move.

Increased rotation speed while attacking.

Info about the collider of an attack (it's damage, it's poise damage, it's flinch strength etc...) are now held inside a DamageColliderInfo struct.

CollisionWithDamageCollider inside PlayerCollisionManager has been divided in smaller functions and now handles parrying too (instead of it being inside the DamageColliderControl script).

Being hit no longer applies a knockback, it now however applies a flinch to the body part being hit in the attack direction.