Parry system rework
Now parries are handled by a sphere collider in front of the player that gets activated when the player parries.

If the damage collider hits the parry collider and the weapon can be parried in that moment, then the player attacking is parried.

If the damage collider collides with a static (a wall for example) and the static is not the ground the player is standing on, then the attack will bounce off.

Fixed a bug causing the player to get stuck in the fall animation when waking up after being knocked out.