Animation system changed, combat with offhand and 2 handing
Animations are now stored in a scriptable object "Animation Data", the events are stored in a list and get "constructed" every time the animation is played, so that the references are correct.
This allows the system to be more modular and less confusing.

The new animation system allows every AnimationData to define a mirrored version of the clip in the same data.

Attacking with the left hand and blocking/ parrying with the right hand is now possible again thanks to the aforementioned system.

Started introduction of two handing weapons and shields.