Animation system change
The AnimationData script now uses Clip Transitions instead of Animation Clips, this allows the script to not specify stuff like animation speed, animation fade in etc... as they are now specified directly in the Clip Transition.

Clip transitions also allow animations to transition smoothly to themselves (attack1 -> attack1...), this was not the case with animation clips.